PWPCC-5 Refactor ProfessionalIdentityHero component structure and add ProfessionalActions and SocialProfileLinks components
