+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import ProfessionalIdentityHero from "./components/ProfessionalIdentityHero";
+import ProfessionalIdentityHero from "./components/hero/ProfessionalIdentityHero";
 
 export default function Home() {
   return <main className="min-h-screen" role="main" id="main-content">
